work so far

+++ main.c
@@ -4,9 +4,9 @@
 // as it requires redundant dynamic memory allocation,
 // while our design is based on constant strings and indices.
 
-    int main() {
-        bool validResult;
-        Node* tree = NULL;
+int main() {
+    bool validResult;
+    Node* tree = NULL;
     char line[MAX_LINE_LEN + 1];
     while (fgets(line, MAX_LINE_LEN, stdin) != NULL) {
         if (strcmp(line, TERMINATION_LINE) == 0) {
@@ -15,7 +15,7 @@
         }
         validResult = true;
         tree = stringToTree(line);
-        float res = calcTree(tree, &validResult);
+        double res = calcTree(tree, &validResult);
         freeTree(tree);
         if (validResult) {
             printf("res = %.2f\n", res);
